requirements.txt
feedparser
Mastodon.py
requests
bs4